+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * Show hyperlinks in dialog windows instead of plain text
  * Move two functions in notifiCat from Tkinter to Qt
  * Fix OpenCV Error when video is at final frame
+ * Don't crash on index error
 
 ## 0.0.1
  * Initial release